Entferne obsolete ungenutzte Definitions und ersetze ageClass-Utility-Methode
