--- src/definitions/age-class-wrapper.ts
@@ -1,6 +1,0 @@
-import { AgeClass } from "./age-class";
-
-export interface AgeClassWrapper {
-  ageClass ?: AgeClass;
-  index: number;
-}
