--- src/definitions/grouped-matches.ts
@@ -1,5 +1,0 @@
-import { Match } from "scraperlib";
-
-export interface GroupedMatches {
-  [id: string]: Match[];
-}
