--- src/definitions/weekend-bucket.ts
@@ -1,5 +1,0 @@
-import { Week } from "./week";
-
-export interface WeekendBucket {
-  [id: number]: Week;
-}
